Fixed the time step when laser is on.
The time step is now tied to the laser frequeny when laser is on.
If the time step is larger than 0.5*T where T is the time period of
laser, the program stops with an error message. If time step is larger
than 0.1*T only a warning message is printed.

 Changes to be committed:
	modified:   src_v4.5.7/ElecLight.cpp
	modified:   src_v4.5.7/parameters.cpp