Removed hard-coded value of hmax for RK4 Integrator
Currently when in 'coherent dynamics' mode and laser on,
the default  hmax (hmax_laser) was hard-coded to be 1 a.u. of time.
Changed it to default to  hmax (hmax_laser) = tstep_laser.

Currently the way hmax for RK4 is toggoled between tstep and
tstep_laser is prone to integrator error. Needs investigation
and fixing.

 Changes to be committed:
	modified:   src_v4.5.7/parameters.cpp